+++ kong/db/schema/metaschema.lua
@@ -79,7 +79,6 @@ local field_schema = {
   { default = { type = "self" }, },
   { abstract = { type = "boolean" }, },
   { generate_admin_api = { type = "boolean" }, },
-  { legacy = { type = "boolean" }, },
   { immutable = { type = "boolean" }, },
   { err = { type = "string" } },
   { encrypted = { type = "boolean" }, },
@@ -337,9 +336,6 @@ local attribute_types = {
     ["integer"] = true,
   },
   uuid = {
-    ["string"] = true,
-  },
-  legacy = {
     ["string"] = true,
   },
   unique = {
@@ -623,12 +619,6 @@ local MetaSchema = Schema.new({
     {
       admin_api_nested_name = {
         type = "string",
-        nilable = true,
-      },
-    },
-    {
-      legacy = {
-        type = "boolean",
         nilable = true,
       },
     },